[HUDI-7441] Move `getWritePartitionPaths` method to common module to decouple hive dependency
